Created a new model and set up REST API
A new model called "Unit" was created to hold our data.
A serializer was added in order to render data as json/xml.
A filter was added to json_view.py in order to sort by id field.
api_urls.py was added in order to display the units.
initial_data.json was used in order to practice migration.
Added "rest_framework"and "django filters"  to settings.py